Add i18n
Create a language picker in the top menu, and switch languages when the
language is changed. English, Dutch and German are supported for now.

The language selection is saved locally in the browser in order to
maintain the selection on page reload.

The labels in the top menu are implemented and translated as a POC, but
other text within needs to be translated still by a translator.