Add view solution page
Create the page where the user can view the solution with a specific
id.

Refactor the SdgColors to Sdgs and have them include the SDG names.

Refactor the solutions to live in a separate file in order to be
defined once and fetch them throughout the application until the
backend is implemented to get the real solutions.